requirements.txt
numpy
scikit-learn
streamlit
tqdm
pillow
kaggle
gdown
plotly
facenet_pytorch
torch  -f https://download.pytorch.org/whl/cu124
torchvision -f https://download.pytorch.org/whl/cu124
